refactor : Move `UserDto.java` file
