+++ src/main/java/kr/kw/matcher/module/article/dto/ArticleDto.java
@@ -2,7 +2,7 @@ package kr.kw.matcher.module.article.dto;
 
 import kr.kw.matcher.module.article.domain.Article;
 import kr.kw.matcher.module.user.domain.User;
-import kr.kw.matcher.module.user.dto.UserDto;
+import kr.kw.matcher.module.user.application.dto.UserDto;
 import lombok.*;
 
 import java.time.LocalDateTime;
